pyproject.toml
[project]
name = "mochi-sync"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.69",
    "pydantic>=2.12",
    "transit-python2>=0.8",
]

[dependency-groups]
dev = ["mypy"]
